backend/requirements.txt
mecab-python3
numpy
unidic
unidic-lite
gensim
uvicorn
fastapi